kv error state support
Whenever kv gets into an error state it stays there until it is reset. This way the user does not need to check the return code of every function unless they really want to, and instead can just check after a "block of serialization". After an error occurs the functions in the kv API immediately return the previous error.