Init video subsystem (#208)
* Initialize the video subsystem

SDL_GetPrimaryDisplay() will return 0 if the video subsystem is not initialized. This is an error code. SDL_GetError() returns:

```
Error: No displays found. Video subsystem has not been initialized
```

* Remove superfluous call

This is unnecessary here and does not do anything.

* Assign the result of init

SDL_Init returns true/false. We can use it to set the flag.