Output all libraries to the same directory
Rather than having all the libraries spread across all the dependency directories, this setting enforces one unified place where all the static and dynamic libraries are generated to.

This simplifies the task of finding the dynamic ones, and/or linking against when not using CMake as the build tool in the actual app.